fix: Filter (color and brand)
