+++ src/App.scss
@@ -7,7 +7,7 @@
         padding-bottom: 0;
     }
 }
-.activeFilterColor {
-    outline-offset: 2px;
-    outline: 3px solid $blue;
-}
+// .activeFilterColor {
+//     outline-offset: 2px;
+//     outline: 3px solid $blue;
+// }
