+++ src/components/Button/Button.js
@@ -4,12 +4,17 @@ import styles from './Button.module.scss';
 
 function Button(props) {
   const {
-    children, handleClick, type, style,
+    children, handleClick, type, className, style,
   } = props;
 
   return (
     /* eslint-disable react/button-has-type */
-    <button type={type} onClick={handleClick} className={`${styles.btn} ${style}`}>
+    <button
+      type={type}
+      onClick={handleClick}
+      className={`${styles.btn} ${className}`}
+      style={style}
+    >
       {children}
     </button>
     /* eslint-enable react/button-has-type */
@@ -20,14 +25,16 @@ Button.propTypes = {
   children: PropTypes.oneOfType([PropTypes.arrayOf(PropTypes.node), PropTypes.node]),
   type: PropTypes.oneOf(['submit', 'button']),
   handleClick: PropTypes.func,
-  style: PropTypes.string,
+  className: PropTypes.string,
+  style: PropTypes.object,
 };
 
 Button.defaultProps = {
   children: '',
   type: 'button',
-  handleClick: () => {},
-  style: '',
+  handleClick: () => { },
+  className: '',
+  style: {},
 };
 
 export default memo(Button);
