+++ src/components/Button/Button.module.scss
@@ -1,10 +1,13 @@
 @import '../../assets/styles/global.module.scss';
 
-.editBtn {
-  font-weight: 600;
-  font-size: 14px;
-  line-height: 21px;
-  border: 2px solid $blue;
-  box-sizing: border-box;
+.btn {
   border-radius: 50px;
 }
+// .editBtn {
+//   font-weight: 600;
+//   font-size: 14px;
+//   line-height: 21px;
+//   border: 2px solid $blue;
+//   box-sizing: border-box;
+//   border-radius: 50px;
+// }
