+++ src/components/FilterCreatorBrand/FilterCreatorBrand.module.scss
@@ -1,0 +1,3 @@
+.brandsCheckbox {
+    margin: 4px 10px 10px 0;
+}
