+++ src/components/FilterCreatorColor/FilterCreatorColor.module.scss
@@ -1,0 +1,29 @@
+@import "../../assets/styles/global.module.scss";
+
+@mixin circleColors($border, $width: 30px, $height: 30px, $margin: 3px, $margin-bottom: 7px, $border-radius: 50px, $overflow: hidden, $cursor: pointer) {
+    width: $width;
+    height: $height;
+    margin: $margin;
+    margin-bottom: $margin-bottom;
+    border-radius: $border-radius;
+    overflow: $overflow;
+    cursor: $cursor;
+    border: $border;
+}
+
+.colorsItems {
+    // width: 30px;
+    // height: 30px;
+    // margin: 3px;
+    // margin-bottom: 7px;
+    // border-radius: 50px;
+    // overflow: hidden;
+    // cursor: pointer;
+    // border: 1px solid $black;
+    @include circleColors(1px solid $black);
+}
+.activeFilterColor {
+    @include circleColors(1px solid $black);
+    outline-offset: 2px;
+    outline: 3px solid $blue;
+}
