+++ src/components/PaginationFilterPage/PaginationFilterPage.js
@@ -2,35 +2,52 @@ import React, { memo, useEffect, useState } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import PropTypes from 'prop-types';
 import ReactPaginate from 'react-paginate';
-import styles from './FilterItems.module.scss';
+import styles from './PaginationFilterPage.module.scss';
 import CardItem from '../CardItem/CardItem';
+import {
+  setFilterPaginationPage,
+} from '../../store/actionCreators/filterAC';
 
-function FilterItems(props) {
+function PaginationFilterPage(props) {
+  const dispatch = useDispatch();
   const { filterProducts, itemsPerPage } = props;
-  const filterByColor = useSelector((state) => state.filter.filterProducts);
+  const filterPaginationPage = useSelector((state) => state.filter.filterPaginationPage);
 
   const [currentItems, setCurrentItems] = useState(null);
   const [pageCount, setPageCount] = useState(0);
   const [itemOffset, setItemOffset] = useState(0);
   const [renderOnPage, setRenderOnPage] = useState(true);
-
+  // useEffect(() => {
+  //   console.log(currentItems);
+  //   console.log(pageCount);
+  //   console.log(itemOffset);
+  // }, [currentItems, pageCount, itemOffset]);
   useEffect(() => {
-    console.log(document.querySelector('a[tabindex="0"]'));
+    console.log(filterPaginationPage);
+    console.log(currentItems);
+  }, [filterPaginationPage]);
+  useEffect(() => {
     setRenderOnPage(false);
     setTimeout(() => {
       setRenderOnPage(true);
     }, 100);
   }, [filterProducts]);
   useEffect(() => {
+    if (filterPaginationPage === 0) {
+      setItemOffset(0);
+    }
     const endOffset = itemOffset + itemsPerPage;
+    console.log(`Loading items from ${itemOffset} to ${endOffset}`);
     if (filterProducts !== null) {
       setCurrentItems(filterProducts.slice(itemOffset, endOffset));
       setPageCount(Math.ceil(filterProducts.length / itemsPerPage));
     }
-  }, [itemOffset, itemsPerPage, filterProducts]);
+  }, [itemOffset, itemsPerPage, filterProducts, filterPaginationPage]);
   const handlePageClick = (event) => {
     const newOffset = (event.selected * itemsPerPage) % filterProducts.length;
     setItemOffset(newOffset);
+    console.log(event.selected);
+    dispatch(setFilterPaginationPage(event.selected));
   };
   return (
     /* eslint-disable no-underscore-dangle */
@@ -52,14 +69,15 @@ function FilterItems(props) {
           </div>
         ))}
       </div>
+      {currentItems?.length === 0 && <h1 className={styles.noItems}>No items</h1>}
       <ReactPaginate
-        forcePage={0} // изначально должно быть равно null (нужно связать с редаксом)
+        forcePage={filterPaginationPage}
+        // изначально должно быть равно null (нужно связать с редаксом)
         // подписать редакс к forcePage и при нажатии на apply filter устанавливать редакс null
         breakLabel="..."
         nextLabel="›"
         previousLabel="‹"
         onPageChange={handlePageClick}
-        pageRangeDisplayed={5}
         pageCount={pageCount}
         renderOnZeroPageCount={null}
         containerClassName={styles.paginateContainer}
@@ -70,14 +88,14 @@ function FilterItems(props) {
   );
 }
 
-FilterItems.propTypes = {
+PaginationFilterPage.propTypes = {
   filterProducts: PropTypes.array,
   itemsPerPage: PropTypes.number,
 };
 
-FilterItems.defaultProps = {
+PaginationFilterPage.defaultProps = {
   filterProducts: [],
   itemsPerPage: 0,
 };
 
-export default memo(FilterItems);
+export default memo(PaginationFilterPage);
