+++ src/pages/AboutProductPage/AboutProductPage.js
@@ -50,7 +50,7 @@ function AboutProductPage() {
         <p>{product.previousPrice}</p>
       </div> */}
       {/* <div className={styles.goBackBtn}>
-        <Button style={styles.aboutProductCartBtn} handleClick={() => navigate(-1)}>Go Back</Button>
+  <Button className={styles.aboutProductCartBtn} handleClick={() => navigate(-1)}>Go Back</Button>
       </div> */}
       <hr className={styles.line} />
       <div className={`${styles.aboutProductTabMenuWrapper} ${styles.aboutProductWrapper}`}>
