+++ src/store/actions/filterActions.js
@@ -9,3 +9,4 @@ export const FILTER_BRAND = 'FILTER_BRAND';
 export const SET_MIN_PRICE_SLIDER_VALUE = 'SET_MIN_PRICE_SLIDER_VALUE';
 export const SET_MAX_PRICE_SLIDER_VALUE = 'SET_MAX_PRICE_SLIDER_VALUE';
 export const NEW_FILTER_PRODUCTS = 'NEW_FILTER_PRODUCTS';
+export const SET_PAGINATION_PAGE = 'SET_PAGINATION_PAGE';
