+++ src/store/reducers/filterReducer.js
@@ -10,6 +10,7 @@ import {
   SET_MIN_PRICE_SLIDER_VALUE,
   SET_MAX_PRICE_SLIDER_VALUE,
   NEW_FILTER_PRODUCTS,
+  SET_PAGINATION_PAGE,
 } from '../actions/filterActions';
 
 const initialValues = {
@@ -21,6 +22,7 @@ const initialValues = {
     min: null,
     max: null,
   },
+  filterPaginationPage: null,
 };
 const filterReducer = (state = initialValues, { type, payload } = {}) => {
   switch (type) {
@@ -50,6 +52,8 @@ const filterReducer = (state = initialValues, { type, payload } = {}) => {
       return { ...state, priceSliderValues: { ...state.priceSliderValues, min: payload } };
     case SET_MAX_PRICE_SLIDER_VALUE:
       return { ...state, priceSliderValues: { ...state.priceSliderValues, max: payload } };
+    case SET_PAGINATION_PAGE:
+      return { ...state, filterPaginationPage: payload };
     default:
       return state;
   }
